requirements.txt
Django>=3.0.8,<3.1.0
djangorestframework>=3.11.0,<3.12.0
django-filter>=2.3.0,<2.4.0
psycopg2-binary==2.8.5
Pillow>=7.2.0,<7.3.0
requests>=2.24.0,<2.25.0
flake8>=3.8.3,<3.9.0
